[WIP] Rework color mode API documentation
Color modes can be difficult to grasp, and I think it might be partially
due to the "happy path" requiring configuration. This attempts to rework
the API to _maybe_ be more straightforward.

Please note that this doesn't yet address any implementation concerns,
but is attempting to document the most desirable API before we
implement. This will be able to be done in a backwards-compatible way.

I'm open to any and all suggestions