Spike out potential TypeScript conversion
This "reworks" `@theme-ui/core` to be written in TypeScript and export
its own type definitions.  Thankfully, `microbundle` has built-in support
for TypeScript so we barely need to do any build system changes. I also
checked in the generated `index.d.ts` file (normally we would not check
this into git) to show what it looks like.