Fix code block line height in Safari
Safari renders a small space between inline-block elements, resulting in a variable height of lines which cause a scroll overflow depending on their width.

On iOS this results in the font size of each token being scaled up individually depending on the width of the token element. On desktop Safari the font size stays consistent and the text causes a line break.