fix(theme-ui): export jsx as createElement
When using the new JSX runtime, Babel will fall back to calling
`createElement` (name not configurable) from the configured
`jsxImportSource` when there is both an explicit `key` prop and
object spread props. This causes a crash when using theme-ui since the
custom JSX pragma is only exported as `jsx`. Reexport the `jsx` pragma
under the name `createElement` to fix this.